Add colorID to Trainer Form and change the datatype for it to an int

+++ Structs/GameDataTypes.cs
@@ -95,7 +95,7 @@ namespace ImpostersOrdeal
         public class Trainer : INamedEntity
         {
             public int trainerTypeID;
-            public byte colorID;
+            public int colorID;
             public byte fightType;
             public int arenaID;
             public int effectID;
